UWexpressions `value` -> `sym` (#222)
* UWexpressions `value` -> `sym`

Renaming the UWexpressions attribute `value` to `sym`. This is
aligned with all sympy objects being called `.sym`.

* Removing `_s` from the Darcy solver.

The gravitational term `s`is now in the DarcyFlowModel. This is because
we use it to build the "flux" for SNES. Update is the ensure this works
with the tests and none of the old implementation remains.

* * Some more `.value' -> '.sym' - for ViscoPlasticFlowModel
* removing warnings