Updates to visuals
Add get_max_radius to mesh because this is sometimes useful
Expressions report their symbol as an ipython display / Markdown as their default display. To see the class structure use expression.view() instead.
MeshVariables report their symbolic form as their default. To see the class / data information use meshVariable.view() instead.
Added a visualisation function to triangulate the coords of a meshVariable in pyvista (2D). The max mesh size is used as the length cutoff for not-drawing / connecting sides. This can be over-ridden using the alpha parameter. Should be good enough for visualisation in most cases. This should not have any side-effects on the new code from thyagi