Adding back a very small smoothing / diffusion term to the projection
In Semi-Lagrange, we have a projection operation to allow us to use different landing discretisations from the original unknown that is being transported. All very well but it is not stable for high gradients. A small smoothing term removes this instability but we do need to check the value for a variety of settings.

Fixing one parallel test to make sure it works with petsc main.