Semi lagrange fixes (#234)
* Add small offset in box point capture

* Clarifications in nodal swarm

* ddt - allowing discontinuous fields for advected points if required

* Trying to fix semi-lagrange scheme

Major bug (totally my fault) in which I failed to account for particles being reordered after sending and receiving. We now store the node and process for each particle which means we can take small substeps to the sample point and then snap back directly to the launch point.

This includes JC's swarm wormhole teleportation code which is required for long trajectories.

I'm turning off all smoothing for projection operators which was introduced to counter numerical problems which turned out to be related to this bug.

Also, added the possibility to use a sample swarm based on a discontinuous underlying variable. This gives a good spread of points and means that corners / facets where particles may be lost are avoided by particle paths.

* Fix regression / testing

* Fixing one broken test

There are (value related) failures in the advection tests. Of course, advection tests are a bloody nightmare since advection schemes are problematic to validate.

Need to understand this before merging PR

* Julian's review comments

posixpath - not sure why this import is present
epsilon in boxmesh return-coords-... - no need to complicate things

* Make substepping time-criterion consistent with timestep definition

* More minor fixes for adv-diffusion

Arise from convection benchmark models.

* A couple of fixes