Implementation of swarm.points
New data structure with callbacks - a sub-class of np.ndarray that synchronises with the petsc data structure when required and calls swarm migration, proxy updating, etc when values are changed.

+= and other in-place operators are disabled because they are difficult to synchronise properly / efficiently.

Various ways to suspend particle synchronisation are included in the arraay implementation.

This does over-print the swarm.data strategy for accessing points if the access manager is used - so more changes are still needed.